% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{absorbing.lookup}
\alias{absorbing.lookup}
\title{Lookup function, absorbing edge}
\usage{
absorbing.lookup(r.val, c.val, BaseProbabilities)
}
\arguments{
\item{r.val}{Row value}

\item{c.val}{Column value}

\item{BaseProbabilities}{Base dispersal probabilities}
}
\description{
Look up the base probabilities when seeds can fall off the edge of the world
}

