% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{as.num}
\alias{as.num}
\title{As number}
\usage{
as.num(x)
}
\arguments{
\item{x}{the thing to be converted to numeric}
}
\description{
A shortcut to avoid having to always write out as.numeric(as.character(x))
}

