% Generated by roxygen2 (4.0.1): do not edit by hand
\name{assign.loop}
\alias{assign.loop}
\title{Assign loop}
\usage{
assign.loop(land.sp, patch.sp.num, p)
}
\arguments{
\item{land.sp}{Species drawn for the landscape}

\item{patch.sp.num}{Number of species to draw for each patch}

\item{p}{The number of cells in a square landscape}
}
\description{
Helper function for randomly assigning species to the landscape
}

