% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{assign.spp.v2}
\alias{assign.spp.v2}
\title{Assign species to the landscape (v2)}
\usage{
assign.spp.v2(p, in.spp, land.sp.num, patch.sp.num)
}
\arguments{
\item{p}{The number of cells in a square landscape}

\item{in.spp}{SpTraits object}

\item{land.sp.num}{Number of species in the landscape}

\item{patch.sp.num}{Number of species in each patch}
}
\description{
Function to assign species to the landscape randomly (without respect to
response trait diversity.
}

