% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{assign.spp.v3}
\alias{assign.spp.v3}
\title{Assign species to the landscape (v3)}
\usage{
assign.spp.v3(p, loc.rtd, land.rtd, land.sp.num, patch.sp.num, unq.rtds,
  SpTraits)
}
\arguments{
\item{p}{The number of cells in a square landscape}

\item{loc.rtd}{The response trait diversity in each patch}

\item{land.rtd}{The response trait diversity in the landscape}

\item{land.sp.num}{The number of species in the landscape}

\item{patch.sp.num}{The number of species in each patch}

\item{unq.rtds}{unq.rtds}

\item{SpTraits}{Species trait information from species file}
}
\value{
Returns a list of species locations in the landscape
}
\description{
Assign species to the landscape, meeting certain response trait diveristy
constraints.
}
\note{
This is not an optimal implementation, and could benefit from some
optimization.  This is especially a problem when there is only one or a few
valid configurations of species.  This is because species are chosen and
removed randomly, so it takes a long time (if ever) to find the valid
configuration(s).
}

