% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boost.quant.reg}
\alias{boost.quant.reg}
\title{Implement boosted quantile regression}
\usage{
boost.quant.reg(bqr.out, RunID, d2, bqr.vars, quantiles,
  NumberOfTrees = 10000)
}
\arguments{
\item{bqr.out}{bqr.out}

\item{RunID}{The run id}

\item{d2}{A data frame}

\item{bqr.vars}{Variables to include in boosted quantile regression}

\item{quantiles}{The quantiles to evaluate at}

\item{NumberOfTrees}{Number of trees to use in Boosted Quantile Regression.
10,000 is suggested as a default. #DD# Look up source from main MS.}
}
\description{
Implement boosted quantile regression
}

