% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc.delta.metrics}
\alias{calc.delta.metrics}
\title{Calculate metric change}
\usage{
calc.delta.metrics(Results, out.metrics, scale.vec)
}
\arguments{
\item{Results}{The results dataframe to store the computed changes}

\item{out.metrics}{The fields for which change needs to be calculated}

\item{scale.vec}{A vector of indicators}
}
\description{
Calculate changes in metrics from beginning of simulation
E.g., change in species richness, or change in functional diversity
}

