% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{calc.delta.metrics.v2}
\alias{calc.delta.metrics.v2}
\title{Calculate metric change}
\usage{
calc.delta.metrics.v2(Results, out.metrics, sv, num.tp)
}
\arguments{
\item{Results}{A dataframe containing the results to be output}

\item{out.metrics}{A list of the fields for which change should be calculated}

\item{sv}{an indicator}

\item{num.tp}{An indicator for the timepoints}
}
\description{
Calculate changes in metrics from beginning of simulation
E.g., change in species richness, or change in functional diversity
}

