% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc.fd}
\alias{calc.fd}
\title{Calculate model functional diversity}
\usage{
calc.fd(Results, SpTraits, abund.mat, spe, resolution, log.trans, sv, num.tp)
}
\arguments{
\item{Results}{The results file to be updated}

\item{SpTraits}{A dataframe containing the Species Trait values}

\item{abund.mat}{A matrix giving abundances for each species and cell}

\item{spe}{The total number of species in the regional species pool}

\item{resolution}{The number of decimal places to round functional traits}

\item{log.trans}{Whether or not to log-transform functional traits (this may be broken)}

\item{sv}{An indicator parameter}

\item{num.tp}{An indicator to assist in updating the correct row of the Results}
}
\description{
This function calculates functional diversity at the end of the model
simulation and updates the Results dataframe.
}
\note{
The intention is to add an option to allow this to be calculated at
intermediate model steps
}

