% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{calc.metrics}
\alias{calc.metrics}
\title{Calculate output metrics}
\usage{
calc.metrics(Results, timepoint, n0.lst, IAdults, spe, p, SpTraits)
}
\arguments{
\item{Results}{A dataframe to contain the results}

\item{timepoint}{The timepoint (initial, final, etc.) for which to calculate the metrics}

\item{n0.lst}{A list of the abundances of each stage for each species}

\item{IAdults}{An index of which stages correspond to adults}

\item{spe}{The total number of species in the species pool}

\item{p}{The number of cells in the landscape}

\item{SpTraits}{A dataframe containing species traits}
}
\description{
Calculates output metrics for the simulation.
E.g., species richness, beta diversity, biomass, and functional diversity
}

