% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc.metrics.v2}
\alias{calc.metrics.v2}
\title{Calculate output metrics}
\usage{
calc.metrics.v2(Results, timepoint, n0.lst, IAdults, spe, p, SpTraits,
  scale.cells, sv, num.tp)
}
\arguments{
\item{Results}{A dataframe to contain the results}

\item{timepoint}{The timpoint being calculated}

\item{n0.lst}{A list of abundances of each stage for each species}

\item{IAdults}{An index indicating which entries correspond to adults}

\item{spe}{The total number of species in the species pool}

\item{p}{The number of cells in a square landscape}

\item{SpTraits}{A dataframe containing species traits}

\item{scale.cells}{A list of cells to restrict the calculation to.}

\item{sv}{An indicator}

\item{num.tp}{An indicator for the timepoint}
}
\description{
Calculates output metrics for the simulation.
E.g., species richness, beta diversity, biomass, and functional diversity
}

