% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculate.traits}
\alias{calculate.traits}
\title{Calculate traits (stochastic plants)}
\usage{
calculate.traits(sp.vals, these.env.cond, sp.out.file, env.file,
  do.seed.clone.override, do.thinning.override)
}
\arguments{
\item{sp.vals}{Input species parameters. See details at \link[spatialdemography]{sp.vals}}

\item{these.env.cond}{Input environmental settings. See details at \link[spatialdemography]{env.param.file}}

\item{sp.out.file}{Location for the output species file}

\item{env.file}{Location for the output Environmental Layers file}

\item{do.seed.clone.override}{An indicator variable to determine whether to override longevity/seed/clone trade-offs}

\item{do.thinning.override}{An inidicator variable to determine whether to override default thinning behavior.}
}
\value{
No value is returned from the function, but a new species file is written and a new environmental layers file is written.
}
\description{
Calculate traits based on logic in stochastic plants MS
}
\note{
Stochastic seed survival probably needs to be stochastic and change at each time step as well.
}
\references{
Gerstenlauer, J.L.K., A.C. Keyel, and K. Wiegand. (in prep.)
Predicting natural selection for life-history traits using stochatic
matrix population models. Target Journal: Journal of Ecology
}

