% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cell.size}
\alias{cell.size}
\title{Cell Size
The size of the cell is determined by the researcher, and needs to be
accounted for when setting up carrying capacities and dispersal
probabilities. The model was originally constructed with a 50 x 50 m
cell size in mind.}
\description{
Cell Size
The size of the cell is determined by the researcher, and needs to be
accounted for when setting up carrying capacities and dispersal
probabilities. The model was originally constructed with a 50 x 50 m
cell size in mind.
}

