% Generated by roxygen2 (4.0.1): do not edit by hand
\name{change.check}
\alias{change.check}
\title{Check for environmental change}
\usage{
change.check(env.c.freq)
}
\arguments{
\item{env.c.freq}{Frequency of environmental change}
}
\description{
Test if there will ever be environmental changes in a model run
}

