% Generated by roxygen2 (4.0.1): do not edit by hand
\name{change.cover.my.env}
\alias{change.cover.my.env}
\title{Update my.env settings}
\usage{
change.cover.my.env(my.env, ev)
}
\arguments{
\item{my.env}{Information about the environmental layers}

\item{ev}{an index for which environmental layer to change}
}
\description{
Function to change my.env due to environmental change
}

