% Generated by roxygen2 (4.0.1): do not edit by hand
\name{change.transition.rates}
\alias{change.transition.rates}
\title{Change Transition Rates}
\usage{
change.transition.rates(my.env, ev)
}
\arguments{
\item{my.env}{Information about the environmental layers}

\item{ev}{An indicator for which environmenal layer to use}
}
\description{
Function to change the transition rates used for landscape change
}

