% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{check.coords}
\alias{check.coords}
\title{Check coordinates}
\usage{
check.coords(x.val, y.val, x.lower.bound, x.upper.bound, y.lower.bound,
  y.upper.bound)
}
\arguments{
\item{x.val}{The X coordinatea}

\item{y.val}{The Y coordinate}

\item{x.lower.bound}{The lower x coordinate of the cell}

\item{x.upper.bound}{The upper x coordinate of the cell}

\item{y.lower.bound}{The lower y coordinate of the cell}

\item{y.upper.bound}{The upper y coordiante of the cell}
}
\description{
Checks if coordinates are within a cell.
}
\note{
Used by make.disp.prob.v2

WARNING Points that fall exactly on a cell's border are dropped.
While this is bad, this is expected to be relatively minor,
as the odds of getting a number that lands exactly on a cell's border is
pretty low, given the number of decimal places.
}

