% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.cum}
\alias{check.cum}
\title{Check that cumulative probabilities are correctly made}
\usage{
check.cum(in.vals, check.probs, check.val)
}
\arguments{
\item{in.vals}{input values}

\item{check.probs}{The probabilities to be checked}

\item{check.val}{The value that the last probability value must have}
}
\description{
Check that cumulative probabilities are correctly made
}

