% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.inputs}
\alias{check.inputs}
\title{Check inputs}
\usage{
check.inputs(tot.sp.num, land.sp.num, patch.sp.num, tot.rtd, land.rtd, loc.rtd,
  p)
}
\arguments{
\item{tot.sp.num}{total number of species in the species pool}

\item{land.sp.num}{Number of species in the landscape}

\item{patch.sp.num}{Total number of species in each patch}

\item{tot.rtd}{total response trait diversity in the species pool}

\item{land.rtd}{total response trait diversity in the landscape}

\item{loc.rtd}{Response trait diversity in each patch}

\item{p}{The number of cells in a square landscape}
}
\description{
Function to check that inputs for random species assignment are valid and
logical.  For example, local patch diversity cannot exceed landscape diversity
}

