% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{check.packages}
\alias{check.packages}
\title{Check for required packages}
\usage{
check.packages(out.metrics, include.copula)
}
\arguments{
\item{out.metrics}{A vector of potential metrics to calculate. Those ending in UTC
will require multirich to calculate, and those ending in FRIC will require
FD to calculate. At some point, I need to add in Petchey & Gaston's FD metric.}

\item{include.copula}{An entry from the settings file. If null or 0, the copula
package is not required.}
}
\description{
This function checks that all required packages needed to run SpatialDemography
are present and usable. Packages to check are:
Matrix, multirich, FD, and copula.
spatialdemography may also use lhs and raster, but these are used in exported
functions that contain their own checks.
Hmisc, quantreg & gbm were formerly listed under suggests, but are currently
not part of the code workflow (i.e. they would require an external call to an internal function)
lattice was formerly listed under suggests, but I don't think it was actually used
so it is no longer listed nor is it checked for.are not integrated into the code.
}

