% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.spp}
\alias{check.spp}
\title{Check species}
\usage{
check.spp(vrs.out)
}
\arguments{
\item{vrs.out}{The vital rates produced during species generation}
}
\description{
Check that the species that were generated are valid and realistic
(e.g. no spontaneous generation)
}

