% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chk.file}
\alias{chk.file}
\title{Check file}
\usage{
chk.file(in.path, in.name, in.format, in.chk, run.name, RunLog)
}
\arguments{
\item{in.path}{in.path}

\item{in.name}{in.name}

\item{in.format}{in.format}

\item{in.chk}{in.chk}

\item{run.name}{Run name}

\item{RunLog}{A log}
}
\description{
Check existence of a file & accuracy of values, based on specified inputs
}

