% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{compile.log.lambdas}
\alias{compile.log.lambdas}
\title{Write log lambdas to file}
\usage{
compile.log.lambdas(in.lists, out.file)
}
\arguments{
\item{in.lists}{One or more lists to be written to file.  Each list should 
have four components: A vector with species headers, a label for the list, a vector with mean log
lambdas, and a vector with sd of log lambda.}

\item{out.file}{The file to be written}
}
\description{
Take several calculations of log lambda, and place them into a single file
with species as rows, and values as columns.
}
\note{
Should change from vec1 and vec2 to take a list of lists as a more robust approach
}

