% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute.diagnostics}
\alias{compute.diagnostics}
\title{Compute Diagnostics}
\usage{
compute.diagnostics(spe, B1.lst, B2.lst, M.lst, P, S, p, outpath.base, vdb.data,
  change.count, run.times, run.lbl, is.first = 0, Gave.Warn = 0)
}
\arguments{
\item{spe}{The number of species in the species pool}

\item{B1.lst}{A list of the first demography matrices for each species}

\item{B2.lst}{A list of the second demography matrices for each species}

\item{M.lst}{A list of the dispersal matrices for each species}

\item{P}{The vec-permutation matrix}

\item{p}{The number of cells in a square landscape}

\item{S}{The number of stages in the matrix model}

\item{outpath.base}{The base path for outputs}

\item{vdb.data}{An indicator for whether or not to create visual debugger data}

\item{change.count}{An indicator for the current environmental change step.}

\item{run.times}{A list of intermediate times during the model run}

\item{run.lbl}{A list of labels for each of the intermediate times}

\item{is.first}{An indicator for whether this is the first time the code is run (for timing purposes)}

\item{Gave.Warn}{An indicator for whether a warning message has already been given. If so, a duplicate warning will NOT be reissued.}
}
\description{
Function to compute basic matrix diagnostics
}
\note{
The diagnostics DO NOT include effects of the carrying capacity or
microsite competition!
WARNING This function requires exactly 4 stages, with the first stage as
the mobile dispersal stage. is stage specific, and requires 4 stages with the first stage as mobile seeds!
}

