% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{compute.funct}
\alias{compute.funct}
\title{Implement standard functions}
\usage{
compute.funct(num.reps, funct, par1, par2 = NA, to.round = 3)
}
\arguments{
\item{num.reps}{The number of values to be created}

\item{funct}{The function to use to create the values}

\item{par1}{See details for chosen function}

\item{par2}{See details for chosen function}

\item{to.round}{Number of decimal places to round output values}
}
\value{
A vector of values with length equal to num.reps
}
\description{
This function implements one of several functions, based on an input code.
}
\details{
Function codes are as follows:
1 = log-normal distribution (par1 = mean, par2 = sd)
2 = normal distribution (par1 = mean, par2 = sd)
3 = constant (par1 = value, par2 = NA)
4 = uniform distribution, continuous values
(par1 = lower bound, par2 = upper bound)
5 = uniform distribution, only integer values
(par1 = lower bound, par2 = upper bound, par1 & 2 must be integers!)
SEE Appendix S1 ODD description for more information
}

