% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute.microsites}
\alias{compute.microsites}
\title{Compute number of available microsites}
\usage{
compute.microsites(microsites.vec, nt1a.lst, spe, ISeedBank, ISeedlings, p)
}
\arguments{
\item{microsites.vec}{A vector giving the number of microsites in each cell.}

\item{nt1a.lst}{A list giving the abundances of each of the four stages for each species}

\item{spe}{The number of species in the regional species pool}

\item{ISeedBank}{An index for which entries for each species in nt1a.lst correspond to stage 1 individuals.}

\item{ISeedlings}{An index for which entries for each species in nt1a.lst correspond to stage 0 individuals.}

\item{p}{Number of cells in the landscape}
}
\description{
Helper function to calculate number of available microsites
}

