% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{Connectance}
\alias{Connectance}
\title{Calculate distances between cells}
\usage{
Connectance(NumPatches, extent)
}
\arguments{
\item{NumPatches}{Number of cells in the landscape, equal to the extent squared}

\item{extent}{Length of one side of the square landscape}
}
\description{
Calculates euclidean distance between all cells in a landscape
}
\details{
Designed with a square landscape in mind, but it will work for other configurations
Basically, cells are added by rows, and if there are not enough cells to complete a row, the row is missing some cells.
}
\author{
Jakob Gerstenlauer, with minor modification by A.C. Keyel
}

