% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{create.stochastic.landscape}
\alias{create.stochastic.landscape}
\title{Create stochastic landscape}
\usage{
create.stochastic.landscape(run.path, my.lyrs, s.lbl, correlation, n, p, fxs,
  fxs.means, fxs.vars)
}
\arguments{
\item{run.path}{Path for the model run}

\item{my.lyrs}{Layers to include in the landscape}

\item{s.lbl}{A label}

\item{correlation}{Desired correlation level (will not be exact)}

\item{n}{Number of samples to generate (1 per time step)}

\item{p}{Number of cells in the landscape}

\item{fxs}{Vector of functions for the copula}

\item{fxs.means}{Vector of means for the copula, in same order as functions}

\item{fxs.vars}{= Vector of variances/sd for the copula, in same order as above}
}
\description{
Create a stochastic landscape for the stochastic plants model using a copula function
In this case, it creates correlated distributions between a log-normal function
(fertility) and a beta distribution (stochastic survival)
}
\note{
Requires copula package.  Cells in the landscape will be generated independently of one another.
}
\references{
Sklar 1959, others? #DD#
}

