% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CreateCompositeDemographyMatrix}
\alias{CreateCompositeDemographyMatrix}
\title{Create a composite demography matrix}
\usage{
CreateCompositeDemographyMatrix(B1.template, B2.template, spe, SpTraits, p, S,
  landscape, landscape.identifiers)
}
\arguments{
\item{B1.template}{A template for creating the B1 matrices}

\item{B2.template}{A template for creating the B2 matrices}

\item{spe}{The number of species in the species pool}

\item{SpTraits}{The traits (i.e. vital rates, response modifiers and dispersal parameters)}

\item{p}{The number of cells in the landscape}

\item{S}{The number of life stages in the matrices}

\item{landscape}{A list object containing the environmental layer values}

\item{landscape.identifiers}{A list of single letter identifiers for each environmental layer}
}
\description{
Creates the demographic matrices used in the matrix model
}
\note{
The original notation used in this function was changed to be consistent & intuitive for ACK.
}
\author{
Jakob Gerstenlauer and Alexander "Sasha" Keyel
}

