% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{CreateCompositeDispersalMatrix}
\alias{CreateCompositeDispersalMatrix}
\title{Create the composite dispersal matrix}
\usage{
CreateCompositeDispersalMatrix(DispersalProbabilities, p, S, M_sub)
}
\arguments{
\item{DispersalProbabilities}{The dispersal probabilities from each cell to
every other cell in the landscape}

\item{p}{Number of cells in the landscape}

\item{S}{Number of life-stages in the model (currently needs to be 4)}

\item{M_sub}{A template matrix of the appropriate dimensions to fill in.}
}
\value{
returns a dispersal matrix for controlling seed dispersal.
}
\description{
Create a composite dispersal matrix for the dispersal portion of the model
}
\details{
This matrix consists of S*S submatrices of dimension p x p describing the
dispersal probability between patches:
Where p is the number of patches, S is the number of stages (here 4).
Only diagonal submatrixes, which describe spatial movement of individuals of
a given stage between patches, can contain non-zero matrix entries.
We assume that only individuals of stage 1 (mobile seeds) move between
patches, therefore only the submatrix M1<-1 contains non-zero off-diagonal
matrix entries!
The off-diagonal entries of M1<-1 describe the probability of seed movement
between patches of the landscape.
The diagonal entries describe the probability that mobile seeds stay in the
local patch, which is 1-u (seed dispersal ratio).
All other diagonal submatrices are identity matrices (ones in the diagonal)
as all individuals of these stages stay put.
0<-0   O     O     O
O     1<-1   O     O
O     O    2<-2    O
O 	   O      O   3<-3
}
\note{
WARNING While S is an input, this function will ONLY work for S = 4.
}
\author{
Jakob Gerstenlauer and Alexander "Sasha" Keyel
}
\references{
Hunter, C.M. & Caswell, H. (2005). The use of the vec-permutation
matrix in spatial matrix population models.
Ecological modelling, 188, 15-21.
}

