% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CreatePermutationMatrix}
\alias{CreatePermutationMatrix}
\title{Create Permutation Matrix}
\usage{
CreatePermutationMatrix(p, S)
}
\arguments{
\item{p}{The number of cells in a square landscape}

\item{S}{The number of stages in the matrix model}
}
\description{
Creates the vec-permutation matrix for use in running the matrix model
}
\author{
Jakob Gerstenlauer
}

