% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cv.setup}
\alias{cv.setup}
\title{Get conditioning layer}
\usage{
cv.setup(target.cond.lyr, landscape, landscape.identifiers)
}
\arguments{
\item{target.cond.lyr}{The layer to use for creating conditional values}

\item{landscape}{A list object containing the environmental layer values for each cell in the landscape}

\item{landscape.identifiers}{A list of single letter identifiers for each environmental layer}
}
\description{
Function to get the values from the conditioning layer
}

