% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disp.lookup.v2}
\alias{disp.lookup.v2}
\title{Look up Dispersal Probabilities}
\usage{
disp.lookup.v2(p, BaseProbabilities, edge.type)
}
\arguments{
\item{p}{The number of cells in a square landscape}

\item{BaseProbabilities}{A set of base dispersal probabilities.}

\item{edge.type}{Either ABSORBING or TORUS. See details.}
}
\description{
Function for looking up dispersal probabilities from a table
}
\details{
The base probabilities will be adjusted for each cell based on the
edge type. In an absorbing edge type, probabilities beyond the landscape will
simply be dropped, and dispersers will be removed (e.g. dispersal off an island)
With a torus edge, dispersers wrap around to the other side of the landscape,
and the dispersal probabilities are adjusted to account for this wrap-around
effect.
}

