% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{do.competition}
\alias{do.competition}
\title{Implement seed competition}
\usage{
do.competition(competitiontype, microsites.vec, nt1a.lst, spe, ISeedBank,
  ISeedlings, p)
}
\arguments{
\item{competitiontype}{The type of competition (see description)}

\item{microsites.vec}{A vector giving the number of microsites in each cell.}

\item{nt1a.lst}{A list giving the abundances of each of the four stages for each species}

\item{spe}{The number of species in the regional species pool}

\item{ISeedBank}{An index for which entries for each species in nt1a.lst correspond to stage 1 individuals.}

\item{ISeedlings}{An index for which entries for each species in nt1a.lst correspond to stage 0 individuals.}

\item{p}{Number of cells in the landscape}
}
\description{
Idea is that seeds will compete for microsites. Currently three types of competition possible:
0 = no competition, 1 = within-species K (excess seeds die), 2 = deterministic
multi-species K (seeds are assigned to microsites proportional to the number of seeds
of each species), and 3 = stochastic multi-species K (seeds are assigned to microsites
via a lottery approach, where each seed has a chance to occupy a microsite.
NOTE: Seeds already in the seedbank occupy microsites, juveniles and adults do not.
WARNING: My implementation of this may be computationally intensive for large numbers of seeds!
}

