% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.copula}
\alias{do.copula}
\title{Function to implement copula}
\usage{
do.copula(correlation, np, fxs, fxs.means, fxs.vars)
}
\arguments{
\item{np}{Number of samples to generate}

\item{correlation}{Desired correlation level (will not be exact)}

\item{fxs}{Vector of functions for the copula}

\item{fxs.means}{Vector of means for the copula, in same order as functions}

\item{fxs.vars}{= Vector of variances/sd for the copula, in same order as above}
}
\description{
Function to implement copula
}
\note{
Currently only works for log-normal and beta distributions.
}
\author{
Jakob Gerstenlauer & Sasha Keyel
}
\references{
Sklar 1959 %DD%
}

