% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.env.change}
\alias{do.env.change}
\title{Implement environmental change}
\usage{
do.env.change(my.env, ev, landscape, landscape.identifiers, p, run.path = NA,
  env.lbl = NA, s.lbl = NA, landscape.dir = NA, lnd.lbl = NA,
  change.count = NA)
}
\arguments{
\item{my.env}{A dataframe containing instructions for how the environment is
to change. See \link[spatialdemography]{env.file} for details.}

\item{ev}{An indicator variable indicating which landscape element is being
modified}

\item{landscape}{A list of environmental layers}

\item{landscape.identifiers}{A list of identifiers for the environmental
layers}

\item{p}{The number of cells in the landscape}

\item{run.path}{The path for this model run (only applicable if c.type == "from.file")}

\item{env.lbl}{A vector containing each environmental layer name in order (only applicable if c.type == "from.file")}

\item{s.lbl}{A label for the scenario being considered(only applicable if c.type == "from.file")}

\item{landscape.dir}{The directory containing the landscape layers for when changes are read in from file}

\item{lnd.lbl}{A scenario specific label identifying the landscape layers to be used.}

\item{change.count}{An indicator for which environmental change is being used.
Note that this will refer to the change PREVIOUS to the current change, as it
is updated at the END of the change process. (only applicable if c.type == "from.file")}
}
\description{
Make environmental changes to a landscape layer according to a set of change
instructions.  The instructions are in the my.env input, which contains the
following elements "env.change.type", "env.change.mag", and "cond.lyr".
"env.change.type" contains the type of change to take place, "env.change.mag"
contains details on the magnitude of those changes, and "cond.lyr" contains
information on what other layers might affect the change.  The change types
depend on whether or not another layer is being used to inform the change
process (if so, it is a conditional landscape element). See
\link[spatialdemography]{env.change.type} for more details.
}

