% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{do.input.rename}
\alias{do.input.rename}
\title{do.input.rename = Helper function to help with renaming}
\usage{
do.input.rename(input.path, to.rename, file.ending)
}
\arguments{
\item{input.path}{Path for input files}

\item{to.rename}{File names to be renamed}

\item{file.ending}{The file ending to add/replace the previous file ending}
}
\description{
Rename input templates with a specified file ending
}

