% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{do.invasion}
\alias{do.invasion}
\title{Implement invasion in MetaComeSE model}
\usage{
do.invasion(nt1a.lst, spe, p, S, SpTraits, num.invaders = 1,
  cells.to.invade = 5, repro.proportion = 1)
}
\arguments{
\item{nt1a.lst}{A list giving the locations of each species in each cell}

\item{spe}{The number of species in the regional species pool}

\item{p}{The number of cells in the landscape}

\item{S}{The number of stages in the population model (= 4)}

\item{SpTraits}{A dataframe containing the species traits}

\item{num.invaders}{The number of species to have invade the landscape}

\item{cells.to.invade}{The number of cells in the landscape to be invaded}

\item{repro.proportion}{The proportion of a species' normal reproduction to
assign to the landscape.  E.g., if a species normally makes 1000 seeds, and
repro.proportion is set to 0.5, 500 seeds will be assigned to each invaded
cell.  Similarly, a repro.proportion of 1.5 would lead to 1500 seeds.}
}
\value{
An updated n0.lst to include the invaders
}
\description{
Manage invasion in the model.  A certain number of species (num.invaders) are
chosen to invade a certain number of cells in the landscape (cells.to.invade)
at a particular propagule pressure (related to the species' reproductive
rates).
}

