% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.qr}
\alias{do.qr}
\title{Quantile regression helper function}
\usage{
do.qr(d2, qr.results, Response, Predictor, this.quantile, counter)
}
\arguments{
\item{d2}{A data frame}

\item{qr.results}{Results}

\item{Response}{Response}

\item{Predictor}{Predictor}

\item{this.quantile}{The quantile of interest}

\item{counter}{A counter}
}
\description{
Helper function to implement the quantile regression
}

