% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.spot.chk}
\alias{do.spot.chk}
\title{Spot check}
\usage{
do.spot.chk(t.spot, test.dat, is.error, RunLog)
}
\arguments{
\item{t.spot}{t.spot}

\item{test.dat}{test.dat}

\item{is.error}{Indicator for whether or not an error occurred}

\item{RunLog}{A run log}
}
\description{
Custom function for doing spot checks
}

