% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.standard.checks}
\alias{do.standard.checks}
\title{Implement checks of data}
\usage{
do.standard.checks(run.name, RunLog, chk.lst, was.vdb = 1)
}
\arguments{
\item{run.name}{Need file path for values to check}

\item{RunLog}{File logging the testing procedure}

\item{chk.lst}{List of inputs for values specific to the run}

\item{was.vdb}{indicator for whether visual debugger data were generated}
}
\description{
Implement checks of data
}
\details{
\itemize{
\item do.standard.checks These are checks that EVERY test run should pass, given the correct inputs \cr
\item do.vdb.checks These are checks that the visual debugger data output correctly \cr
\item do.chg.checks Check that environmental change was implemented correctly \cr
}
}

