% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{do.thinning}
\alias{do.thinning}
\title{Apply thinning law}
\usage{
do.thinning(initial.biomass, final.biomass, m = -4/3)
}
\arguments{
\item{initial.biomass}{Initial biomass}

\item{final.biomass}{Final biomass}

\item{m}{The thinning law, defaults to -4/3}
}
\description{
Calculate a component of transition rates based on a thinning law approach.
Inputs may be vectors of values
}
\details{
The thinning law here affects transition rates
(biomass(t + 1) / biomass (t))** m, where m = -4/3 (reciprocal of the logarithmic self-thinning slope.
The idea is that there are more individuals than can be supported by the environment
and as individuals grow, they interfere with each other's ability to obtain resources.
Some individuals die as a consequence of this competition (typically the smaller
individuals, although size and competition is not directly modeled (hence
assumption #3). \cr
Assumptions: \tabular{ll}{
(1) \tab There are more individuals at time t than can reach time t + 1 \cr
(2) \tab Even aged stand \cr
(3) \tab Survival is probabilistic, and can be applied without knowledge of
individual characteristics and/or identities. \cr
}
}
\note{
Thinning laws were developed in a single-species, even-aged stand context.
Consequently, their application in a multi-species, uneven-age community context
may violate critical assumptions.  See Gerstenlauer et al. for a more
thorough discussion.
}
\references{
Wiegand et al. unpubl. #DD# check if this is published (I think I have it)

Gerstenlauer, J.L.K., A.C. Keyel, and K. Wiegand. (in prep.)
Predicting natural selection for life-history traits using stochatic
matrix population models. Target Journal: Journal of Ecology
}

