% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{do.warn}
\alias{do.warn}
\title{Issue warnings}
\usage{
do.warn(warn.lst1, warn.lst2)
}
\arguments{
\item{warn.lst1}{First warning list}

\item{warn.lst2}{Second warning list}
}
\description{
Function to issue warnings
}

