% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{do.z.transform}
\alias{do.z.transform}
\title{Calculate Z-transform}
\usage{
do.z.transform(d2, in.vars)
}
\arguments{
\item{d2}{A data frame}

\item{in.vars}{input variables}
}
\description{
Calculate Z-transform
}
\author{
Jakob Gerstenlauer, converted to function by Sasha Keyel
}

