% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drop.partial}
\alias{drop.partial}
\title{Drop fractional individuals}
\usage{
drop.partial(x)
}
\arguments{
\item{x}{A number >= 0 (corresponds to abundance)}
}
\value{
Either the original number (if x >= 1) or 0 (x < 1).
}
\description{
Abundances >= 1 remain as they were
Abundances < 1 are recoded as 0
Negative inputs are not allowed.
}

