% Generated by roxygen2 (4.0.1): do not edit by hand
\name{env.file}
\alias{env.file}
\title{Environmental Layers File}
\description{
Environmental settings.  There should be a row for every environmental layer
included in the model.  In the case of pre-generated, unchanging landscapes,
only the env.lbl, landscape.identifiers, and env.change.freq fields are important.
}
\details{
\tabular{lll}{
Input \tab Example \tab Description\cr
env.lbl \tab mowing \tab A label for the environmental layer. \cr
landscape.identifiers \tab M \tab A single letter identifier for each environmental layer.\cr
env.type \tab cover \tab A code for the type of environmental layer. See \link[spatialdemography]{env.type} for more details.\cr
cover.levels \tab 0.75;0.25 \tab Cover levels for each value of the environmental variable (separated by a semi-colon). Only applicable if env.type is set to cover. See \link[spatialdemography]{env.type} for more details. \cr
param1 \tab 0;1 \tab A parameter guiding the construction of the environ-mental layer. See \link[spatialdemography]{env.type} for more details.\cr
param2 \tab NA \tab A second parameter guiding the construction of the en-vironmental layer. See \link[spatialdemography]{env.type} for more details.\cr
cond.var \tab landcover \tab If NA, the current landscape element will be produced independently.  Otherwise, this generation of this environmental layer will depend on the environmental layer specified here. In theory multiple environmental layers could be specified.  In practice, this functionality is not fully supported.\cr
cond.vals \tab 1 \tab Indicates which values in the other layer should be con-sidered potentially suitable (this layer will be assigned values of 0 for all other values of the conditioning layer). If cond.var is NA, this should also be NA.\cr
env.change.freq \tab 10 \tab Indicates how often the environment should change.  A 0 turns off environmental change. There is currently no way to get an irregular change interval, or to stop environmental change mid-scenario based on this parameter.\cr
env.change.type \tab swap \tab Indicates what type of environmental change should take place. See \link[spatialdemography]{env.change.type} for more details. \cr
env.change.mag \tab NA \tab Describes the magnitude of the environmental change. See \link[spatialdemography]{env.change.type} for more details.\cr
}
}

