% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{ethreshold}
\alias{ethreshold}
\title{Simple function to apply an extinction threshold}
\usage{
ethreshold(in_value, e_threshold)
}
\arguments{
\item{in_value}{value to check against threshold}

\item{e_threshold}{Threshold. Values less than this will be rounded to zero.}
}
\value{
A value modified according to the threshold
}
\description{
Simple function to apply an extinction threshold
}

