% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract.message}
\alias{extract.message}
\title{Give extraction message}
\usage{
extract.message(abund.vec, Missing.Values, Stage, TimeStep)
}
\arguments{
\item{abund.vec}{A vector of abundances}

\item{Missing.Values}{An indicator for whether or not there were missing values (prevents repeat messages)}

\item{Stage}{The stage where the first missing value was detected}

\item{TimeStep}{The timestep where the first missing value was detected.}
}
\description{
Give message if extraction failed for a lifestage and timestep (but only if a similar warning has not already been given.
}

