% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract.scale.cells}
\alias{extract.scale.cells}
\title{Extract Scale Cells}
\usage{
extract.scale.cells(scale.cells.lst, sv, S)
}
\arguments{
\item{scale.cells.lst}{A list of cells to be kept}

\item{sv}{An indicator}

\item{S}{The number of life stages in the model}
}
\description{
Extract the indices of cells to keep (indicated by T) and those to exclude (indicated by F).
Basically, the vectors that will be extracted from will have S elements for
each cell. Therefore the initial scale.cells entry needs to be expanded to accomodate this.
}

