% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{find.disp}
\alias{find.disp}
\title{Find Displacement}
\usage{
find.disp(i.pos, disp, d.pe, d.ne, n)
}
\arguments{
\item{i.pos}{Position of the cell in the landscape}

\item{disp}{Total displacement to be applied}

\item{d.pe}{Distance to the positive edge (top or right)}

\item{d.ne}{Distance to the negative edge (bottom or left)}

\item{n}{Dimension of the landscape}
}
\description{
Function to find the displacement for a torus landscape, using wrap around in the landscape
}

