% Generated by roxygen2 (4.0.1): do not edit by hand
\name{force.hibernate}
\alias{force.hibernate}
\title{Hibernate}
\usage{
force.hibernate()
}
\description{
Hibernate the computer.
}
\details{
Can be used following simulation runs to shut-down the computer.
Hibernate is chosen over shutdown, as this allows direct access to the
current R session (rather than requiring reloading a saved session, where you
can no longer see the commands & have to deal with the changed workspace.
Note that the force option is used, which will cause the computer to
hibernate even if other programs are running, or if the computer is locked.
}

