% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ForEach}
\alias{ForEach}
\title{ForEach}
\usage{
ForEach(DataFrame, header, FunctionName, Appendix = "")
}
\arguments{
\item{DataFrame}{The data frame of reference}

\item{header}{A vector of type character containing the names of all variables}

\item{FunctionName}{The name of function which is used to transform all variables}

\item{Appendix}{appendix to new variable name}
}
\description{
This function implements a simple algorithm transforming all variables that are part of the header vector with the function in FunctionName.
Note that all variables are overwritten! by default
The <<- operator creates objects in the global environment which is necessary here!
}
\author{
Jakob L.K. Gerstenlauer
}

