% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{gen.spp}
\alias{gen.spp}
\title{Generate species}
\usage{
gen.spp(sp.base.file, sp.resp.instr.file, sp.out.file, tot.sp.num)
}
\arguments{
\item{sp.base.file}{An input file containing base vital rates.
See Appendix S1 (ODD protocol) for file setup & details}

\item{sp.resp.instr.file}{A file containing details on the species response
traits, and how they should be distributed.  Currently these are independent
of the underlying base vital rates.}

\item{sp.out.file}{The species file to be created}

\item{tot.sp.num}{The total number of species to be generated}
}
\description{
This function generates species based on the instructions given in two input
files.
}
\details{
sp.base.file contains the base vital rates and how they should be distributed
in the regional species pool, while resp.traits contains information on 
species response traits.  This creates a regional species pool that contains
a realization of a defined distribution of base vital rates and response
traits.
}
\note{
Currently does not support creation of constrained species traits,
although the intention is to add that functionality in future versions.
WARNING There appears to be a bug in R (3.1.0) that is affecting the way
some column headers are being displayed during troubleshooting. (this will
not affect function users and only matters for those modifying the functions)
%% Notes on function development in 2014_05_MC.docx file,\
%% under 2014-05-06 (the second one)
}

