% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{get.adj.pos}
\alias{get.adj.pos}
\title{Get adjusted positions}
\usage{
get.adj.pos(row.indicator, cell, S, p)
}
\arguments{
\item{row.indicator}{An indicator for the row}

\item{cell}{The target cell}

\item{S}{The number of stages in the matrix model}

\item{p}{The number of cells in a square landscape}
}
\description{
Get positions for adjusting the B1 matrix and for assigning to the B2 matrix
}

