% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.base.env.lyr}
\alias{get.base.env.lyr}
\title{Create base environmental layer}
\usage{
get.base.env.lyr(env.info, p, to.round = 1)
}
\arguments{
\item{env.info}{A list with three parts: one function and two parameters}

\item{p}{The number of cells in the landscape}

\item{to.round}{Number of decimal places in the resulting environmental layer}
}
\value{
returns a created environmental layer (a vector)
}
\description{
Create a base environmental layer that can then be arranged into different
configurations.
%%Add details to explain different options & link to where functions are defined.
}
\note{
Note that the created environmetnal layer is converted to a matrix by
rows, rather than by columns.
}

