% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{get.cond.env.lyr}
\alias{get.cond.env.lyr}
\title{Create conditional environmental layer}
\usage{
get.cond.env.lyr(env.info, p, other.lyrs, other.lyrs.info)
}
\arguments{
\item{env.info}{A list with three parts: one function code and two parameters}

\item{p}{The number of cells in the landscape}

\item{other.lyrs}{The other layer (or layers) to take into account}

\item{other.lyrs.info}{The values in the other layer that represent suitable
habitat}
}
\value{
An environmental layer with a value for each cell in the landscape
(as a vector)
}
\description{
Generates an environmental layer, taking into account the values of another
environmental layer.  The new layer will be assigned only for particular
values of the other layer, for all other values, the new layer will have the
value of zero.
}
\note{
WARNING: Not tested for conditioning on more than one layer.  It should work, but needs testing
}

