% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{get.cors}
\alias{get.cors}
\title{Calculatae spearman correlations}
\usage{
get.cors(qr.cor.out, RunID, d2, cor.vars)
}
\arguments{
\item{qr.cor.out}{qr.cor.out}

\item{RunID}{An ID}

\item{d2}{A data frame}

\item{cor.vars}{cor.vars}
}
\description{
Calculate spearman correlations for selected variables
}
\author{
Jakob Gerstenlauer and Sasha Keyel
}

