% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{get.disp.prob}
\alias{get.disp.prob}
\title{Get dispersal probabilities}
\usage{
get.disp.prob(sp, SpTraits, p, distances, outpath, num.sim, edge.type)
}
\arguments{
\item{sp}{The species of interest (numeric)}

\item{SpTraits}{A dataframe holding information about the species}

\item{p}{the number of cells in the landscape}

\item{distances}{The intercell distances for every cell in the landscape}

\item{outpath}{The path where a dispersal probability table is, or should be written to.}

\item{num.sim}{The number of draws to use in simulating dispersal}

\item{edge.type}{The type of landscape edge (i.e. ABSORBING or TORUS)}
}
\description{
Get dispersal probabilities
}

