% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.env.lyr}
\alias{get.env.lyr}
\title{Create environmental layer configuration}
\usage{
get.env.lyr(base.lyr)
}
\arguments{
\item{base.lyr}{An environmental layer (as a vector)}
}
\value{
An environmental layer (as a vector)
}
\description{
Takes the layer generated by get.base.env.lyr and draws a sample from it
to create a new environmental layer configuration
}
\note{
Note that the vectors are converted to a matrix by rows, rather than by
columns
}

