% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{get.log.lambda}
\alias{get.log.lambda}
\title{Calculate log lambda and dispersion of log lambda}
\usage{
get.log.lambda(in.file, lbl)
}
\arguments{
\item{in.file}{an input file}

\item{lbl}{a label}
}
\description{
Like the title says. %%DD%%
}

