% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.mod}
\alias{get.mod}
\title{Get Modifier}
\usage{
get.mod(env.val, opt.val, funct, par1, par2)
}
\arguments{
\item{env.val}{The value from a focal environmental layer for a given cell}

\item{opt.val}{The target value for the focal species}

\item{funct}{A function describing how to determine the modifier based
on env.val and opt.val.  A list of possible functions is given in #LL# LOOK UP}

\item{par1}{The first input parameter to the function.}

\item{par2}{The second input parameter to the function}
}
\value{
Returns a modifier between zero and one that can be multiplied with
the base vital rate.
}
\description{
Purpose is to calculate vital rate modifiers based on input functions.
}
\author{
Alexander "Sasha" Keyel
}

