% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{get.positions}
\alias{get.positions}
\title{get positions}
\usage{
get.positions(x, cell, S, p)
}
\arguments{
\item{x}{The vital rate in need of a position}

\item{cell}{cell}

\item{S}{The number of stages in the matrix model}

\item{p}{The number of cells in a square landscape}
}
\description{
Get positions for assigning vital rates to the appropriate matrix locations
}

