% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{get.x}
\alias{get.x}
\title{Get x/y coordinate}
\usage{
get.x(a.distance, an.angle, angle.units = "RADIANS")
}
\arguments{
\item{a.distance}{The input distance}

\item{an.angle}{The input angle}

\item{angle.units}{The units for the angle. Must be set to "RADIANS".}
}
\description{
Functions for getting an x (or y) coordinate from angle & distance
}

