% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{getfd}
\alias{getfd}
\title{Calculate functional diversity (helper function to calc.fd)}
\usage{
getfd(Results, SP.mod, abund.mat, resolution, log.trans, sv, num.tp)
}
\arguments{
\item{Results}{The results file to be updated}

\item{SP.mod}{Species Traits, but with transformations and changes in format}

\item{abund.mat}{A matrix giving abundances for each species and cell}

\item{resolution}{The number of decimal places to round functional traits}

\item{log.trans}{Whether or not to log-transform functional traits (this may be broken)}

\item{sv}{An indicator parameter}

\item{num.tp}{An indicator to assist in updating the correct row of the Results}
}
\description{
Calculate functional diversity for calc.fd & place values in Results dataframe
Choose what metrics to calculate based on what is set up in Results.
Calculate appropriate functional trait diversity
Calculate for both functional traits & just for response traits
}

