% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{IA.recode}
\alias{IA.recode}
\title{Convert double T's to T, otherwise set to F}
\usage{
IA.recode(x)
}
\arguments{
\item{x}{the input to be recoded}
}
\description{
I'm sure this exists as a function!
}

