% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{id.extract}
\alias{id.extract}
\title{ID Extract}
\usage{
id.extract(Stage, TimeStep, sub.dat)
}
\arguments{
\item{Stage}{The lifestage to be extracted}

\item{TimeStep}{The timestep to be extracted}

\item{sub.dat}{The dataset containing the data to extract from.}
}
\description{
Simple extraction function to streamline code
}

