% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{import.env.lyr}
\alias{import.env.lyr}
\title{Import environmental layer}
\usage{
import.env.lyr(lyr.lst, lyr.format.lst, outfile)
}
\arguments{
\item{lyr.lst}{A list of environmental layer values (one for each change step)}

\item{lyr.format.lst}{A list specifying the format for each set of environmental layer values in lyr.lst}

\item{outfile}{The landscape file to be generated.}
}
\description{
Takes an existing environmental layer in the format of an R matrix or raster and
converts to the required input format for SpatialDemography. This may be used for 
multiple change steps for a single environmental layer (e.g., if there is
mowing data for multiple years)
}

