% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initial.conditions.file}
\alias{initial.conditions.file}
\title{Initial Conditions File}
\description{
This file contains the initial model conditions for one or more scenarios.
Each row in this file corresponds to a scenario.
}
\details{
\tabular{ll}{
Input \tab Description \cr
Scenario \tab A label for each scenario \cr
ModelName \tab A name for each scenario \cr
Extent \tab The number of cells on one edge of the landscape (landscapes are square, and contain Extent2 cells). \cr
Edge.Type \tab The type of edge to use for the simulation.  Options are TORUS and ABSORBING. \cr
MaxTime \tab The number of time steps to include in the model. \cr
K_g \tab The adult biomass carrying capacity, in grams.  To get K in terms of abun-dance, divide K_g by the biomass of each species. \cr
multi.species.K \tab An variable indicating carrying capacity type.  When set to 0 the carrying capacity is calculated for each species separately, when set to 1, the carry-ing capacity is applied to the biomass of ALL species combined. \cr
competitiontype \tab A variable indicating competition type.  When 0, stage 0 individuals do not compete for microsites.  When 1, stage 0 individuals compete with other individuals of the same species for microsites, when 2, stage 0 individuals compete deterministically with all other stage 0 individuals for microsites. When 3, species compete with all other species stochastically (lottery model). \cr
microsites \tab Total number of microsites that can be occupied. \cr
invasion \tab An indicator for whether or not invasion should take place.  No invasion takes place when 0, otherwise, the variable sets the invasion frequency in number of timesteps (e.g., 1 would have invasion at every time, 2 would have invasion at every 2nd timestep). \cr
num.invaders \tab The number of invaders to be added to a cell. \cr
cells.to.invade \tab The number of cells in the landscape to be invaded. \cr
invader.repro.potential \tab the propagule pressure of the invaders (based on their reproductive potential). \cr
starting.biomass \tab The starting biomass to be initialized in each cell.  Will be divided evenly among non-dispersing life stages.  Only used/needed if initial.n == 2 in \link[spatialdemography]{settings.file}.\cr
num.adults \tab The number of adults to be initialized in each cell.  Only applies when automatically assigning species to the landscape. \cr
Tot.sp.rich \tab The number of species to include in the regional species pool.  Species in the regional species pool can be assigned to the landscape, and are available to invade the landscape.  When generating species, this will set the number of species to create; when reading in species from file, this will serve as a check that the appropriate number is generated. this species richness will reflect the regional species pool, and can be used for assignment to the landscape or for subsequent invasion. \cr
Land.sp.rich \tab The number of species initially in the landscape.  Only relevant when generating species automatically \cr
Loc.sp.rich \tab The number of species assigned per cell. (currently the same number for automatic generation). \cr
Tot.rtd \tab Total species pool response trait diversity.  This includes dispersal traits as well as response traits from the species responses file. \cr
Land.rtd \tab Total landscape response trait diversity. \cr
Loc.rtd \tab Total local response trait diversity. \cr
}
}

