% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.ergodic}
\alias{is.ergodic}
\title{2. TEST ERGODICITY OF A MATRIX}
\usage{
is.ergodic(x)
}
\arguments{
\item{x}{a square nonnegative matrix}
}
\description{
This function will test the ergodicity of a matrix based upon whether its dominant
left eigenvector is positive or not.  A reducible-ergodic matrix has a positive
dominant left eigenvector, where a reducible-nonergodic matrix will contain zeroes
in its dominant left eigenvector.  An irreducible matrix is always ergodic.
NOTE: this function can prove problematic for imprimitive, irreducible matrices or
reducible matrices with imprimitive, irreducible blocks on the diagonal.  The
'dominant' left eigenvector chosen by R is that with the largest absolute value
(including both real and imaginary components), so that sometimes it chooses a dominant
with imaginary components.  In addition, where a reducible matrix in block-permuted form
contains one or more irreducible, imprimitive blocks on the diagonal, values of zero in
the dominant left eigenvector may actually be calculated by R as small numbers, approximate
to zero.  Hence, for such matrices, R may return a spurious result of the matrix being ergodic
when in fact it is nonergodic.  For this reason, the 'dominant' left eigenvector is also returned
by the function so that it may be checked by eye: if it contains non-zero imaginary components,
or if it contains small numbers that may approximate to zero then further diagnosis may be
required.  The 'blockmatrix' and 'is.primitive' functions can help to
diagnose whether diagonal blocks are imprimitive or not.
}
\author{
Iain Stott, 29.1.2010
}

