% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.irreducible}
\alias{is.irreducible}
\title{1. TEST REDUCIBILITY OF THE MATRIX}
\usage{
is.irreducible(x)
}
\arguments{
\item{x}{a square nonnegative matrix}
}
\description{
This function tests whether a matrix is reducible or irreducible.
IT WILL ONLY RELIABLY WORK FOR NONNEGATIVE MATRICES (e.g. PPMs).
It works on the basis that for any matrix A, (I+A)^(s-1) where I
is the identity matrix and s is equal to the order of the matrix
is positive (Caswell 2001).  The function returns the condition
"TRUE: Irreducible" or "FALSE: Reducible".
}
\author{
Iain Stott, 9.1.2009
}

