% Generated by roxygen2 (4.0.1): do not edit by hand
\name{landscape.files}
\alias{landscape.files}
\title{Landscape Files}
\description{
The landscape files are optional, and only are needed if a specific landscape
configuration is desired.  In this case, there should be a file for each
environmental layer in the landscape (with each file name matching an env.lbl
field  in the Environmental_layers file plus an _0.csv, e.g., for landcover,
the file name would be landcover_0.csv).  Within each file the values for the
environmental layer should be present in comma-separated format, with the
number of rows equaling the number of columns equaling the landscape extent
specified in the Initial_conditions file.  There should be no column or row
headers.
}
\details{
During the model run, if there is environmental change, new landscapes can be
generated.  These will have a number appended after the file name, and the numbers
will be linked to correpsonding timesteps in the \link[spatialdemography]{change.lookup}.
}

