% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdemography.r
\name{landscape.object}
\alias{landscape.object}
\title{The landscape object}
\description{
An alternative to inputing a landscape directory is to construct the
landscape in R. To do this, there must be a list, containing a vector for
each environmental variable (in the same order as they are listed in the 
environmental layers file, see \link[spatialdemography]{env.file}). The 
vector should contain the values for each cell in a square landscape, given
by rows (i.e. starting at the top left and going to top right, then down
to next row, and so forth, until the bottom right corner is reached). Note
that the data are given in vector format, not as a square matrix.
}

