% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.seed}
\alias{load.seed}
\title{Load Seed}
\usage{
load.seed(run.id, seed.id, testing)
}
\arguments{
\item{run.id}{Identifies which test run is being run}

\item{seed.id}{Identifies the location where the seed needs saving}

\item{testing}{TRUE/FALSE Indicator for whether a test is being run}
}
\description{
Simple function to load a seed token (and ensure that you don't break the normal code while doing so!
WARNING: Output from this function needs to be assigned to .Random.seed with the global assignment operator <<-
OTHERWISE the assignment will be local to the FUNCTION ONLY and will not properly reset the random seed!
}

