% Generated by roxygen2 (4.0.1): do not edit by hand
\name{local.transition.matrices}
\alias{local.transition.matrices}
\title{Local Transition Matrices}
\description{
These files give the local transition matrices for each species for each cell
in the landscape for each set of environmental conditions.  See Appendix S1
for more details.
}

