% Generated by roxygen2 (4.0.1): do not edit by hand
\name{locations.file}
\alias{locations.file}
\title{Initial Species Locations File}
\description{
An Initial Species Locations File is also optional, and should only be
included if particular species starting locations are desired.
There are two formats for this file. In the old format, there should be one
row for each cell in the landscape.  In each row, the species that should be
present in the corresponding cell should be listed, separated by commas
All species will be initialized with the same number of adults as specified
in the num.adults field in the Initial Conditions File
\link[spatialdemography]{initial.conditions.file}). The file should not
include header information. In the new format, the file should be in the
format of the SpeciesData.csv file generated by the model. The header should
include (in this order): LifeStage, Species, TimeStep, Cells (one column for
each cell in the landscape) Only locations from a single (user-specified)
timestep will be extracted.
}

