% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{make.cum}
\alias{make.cum}
\title{Make cumulative probabilities}
\usage{
make.cum(in.transitions)
}
\arguments{
\item{in.transitions}{a matrix of transition probabilities, with rows corresponding to the transition probabilities for each value}
}
\description{
Converts from a series of individual transition probabilities to a cumulative transition probability
}

