% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{make.disp.prob.v2}
\alias{make.disp.prob.v2}
\title{Generate dispersal probabilities}
\usage{
make.disp.prob.v2(funct, par1, par2, outpath, num.recs = 1e+05,
  max.disp.extent = 101, cut.off = 1e-05)
}
\arguments{
\item{funct}{Chose a function for modeling dispersal (e.g., log-normal).
Function codes are 1: log-normal, 2: normal, 3: constant, 4: uniform,
5: uniform, rounded to nearest integer.}

\item{par1}{First function parameter.  For log-normal, this should be the
mean dispersal distance in cell units.  For example, if mean dispersal for a
species was 25 m, and you intend to use a 50 m cell size, you would enter
25 / 50 = 0.5 here.  Note that this is NOT log-transformed}

\item{par2}{Second function parameter.  For log-normal, this is the dispersal
standard deviation in log-transformed  cell units.}

\item{outpath}{outpath specifies where to put the newly generated dispersal
probability table}

\item{num.recs}{The number of draws used in the simulation.  Ideally this
will be a large number (e.g., 1,000,000), but the function will run slowly
for numbers >10,000.}

\item{max.disp.extent}{The maximum dispersal distance allowed for calculation
of probabilities.  With the log-normal function and a high enough standard
deviation, you can get VERY long distance dispersal.  This option sets a
cut-point to improve computational efficiency.}

\item{cut.off}{The minimum allowed dispersal probability.  Probabilities
less than this value will be rounded to 0.  This is intended to improve
computational efficiency.}
}
\value{
No value is returned, the function is run for its side-effects
(creation of a dispersal probability table.)
}
\description{
Calculates dispersal probabilities (point to area) via simulation based on
It creates a table with dispersal probabilities to different cells compatible with the SpatialDemography model.
an input mean dispersal distance (not log transformed) and dispersal standard deviation (log-transformed).
}
\details{
Point to area dispersal assumes that dispersers originate in the center of
the cell, and determines dispersal probabilities using the entire area of
potential destination cells. (see Chipperfield et al. 2011 for validation of
this approach). Unlike Chipperfield et al. 2011, this function is not based
on integrals, but uses a simulation approach to approximate the dispersal
probabilities.(not as mathematically sophisticated, but much simpler to
program, and with large enough sample sizes, the differences are minimal.)
}
\note{
WARNING: If dispersers disperse out of the landscape extent (set by
max.disp.extent), they will still disperse out of the landscape, regardless
of whether a torus or absorbing edge is used.
Only tested for log-normal distribution so far.  Also, it may be faster
to use a GIS implementation for this process.
}
\author{
Alexander "Sasha" Keyel
}
\references{
Chipperfield, J.D., E.P. Holland, C. Dytham, C.D. Thomas,
T. Hovestadt. 2011.  On the approximation of continuous dispersal kernels in
discrete-space models.  Methods in Ecology and Evolution 2: 668-681.
}

