% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.final.graphs}
\alias{make.final.graphs}
\title{Make final graphs}
\usage{
make.final.graphs(in.pdf, ResultsFile, cover.levels, s.lbl)
}
\arguments{
\item{in.pdf}{The file to be generated}

\item{ResultsFile}{The file that results will be written to.}

\item{cover.levels}{The percent cover of each layer type}

\item{s.lbl}{a label}
}
\description{
THIS FUNCTION IS IN PROGRESS
}

