% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{make.species.locs}
\alias{make.species.locs}
\title{Make species locations element}
\usage{
make.species.locs(p)
}
\arguments{
\item{p}{The number of cells in a square landscape}
}
\description{
Helper function for extraction type 2
The initial entry will be labeled "none", so as not to confuse the regex experession.
}

