% Generated by roxygen2 (4.0.1): do not edit by hand
\name{markov.change}
\alias{markov.change}
\title{Markov change}
\usage{
markov.change(l.lyr, my.env, ev, p, cond.lyr = NA, cond.vals = NA)
}
\arguments{
\item{l.lyr}{The current landscape layer values}

\item{my.env}{Information about the environmental layers}

\item{ev}{An indicator for which environmenal layer to use}

\item{p}{The number of cells in the landscape}

\item{cond.lyr}{Another layer to use for conditional setup}

\item{cond.vals}{Values in the other layer to use as suitable}
}
\description{
Function to do environmental change via a markov chain (transition probabilities)
}

