% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matrix.as.vector}
\alias{matrix.as.vector}
\title{matrix.as.vector}
\usage{
matrix.as.vector(in.mat)
}
\arguments{
\item{in.mat}{A matrix}
}
\description{
Converts a matrix to a vector, but does so by rows (I don't know of a pre-existing option for this)
}

