% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{OccSetup}
\alias{OccSetup}
\title{Set up initial population sizes}
\usage{
OccSetup(n0.lst, spe, species.locs, S, n.seed, n.juv, n.adult)
}
\arguments{
\item{n0.lst}{A list of initial population sizes for each species (will be empty on input, will be returned with values)}

\item{spe}{The total number of species in the species pool}

\item{species.locs}{A list of which cells are occupied by each species}

\item{S}{The number of life stages}

\item{n.seed}{The number of seeds to assign to each occupied cell}

\item{n.juv}{The number of juveniles to assign to each occupied cell}

\item{n.adult}{The number of adults to assign to each occupied cell}
}
\description{
Set up initial population sizes in the model
}

