% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdemography.r
\name{overall.matrix.diagnostics}
\alias{overall.matrix.diagnostics}
\title{Overall Matrix Diagnostics}
\description{
This summary file provides the landscape-wide growth rate, irreducibility,
ergodicity, maximum, median, and minimum local growth rates, maximum, minimum,
and median seed imports and exports, and the proportion of seeds dispersing.
(the model says max, min and median, but as dispersal is a species trait, it
the proportion dispersing will not change). \cr
Each row corresponds to a species.
}

