% Generated by roxygen2 (4.0.1): do not edit by hand
\name{patch.assign.v2}
\alias{patch.assign.v2}
\title{Patch level species assignment}
\usage{
patch.assign.v2(rtds.to.use, loc.rtd, sp.subset, patch.sp.num, warn.lst1,
  warn.lst2)
}
\arguments{
\item{rtds.to.use}{rtds.to.use}

\item{loc.rtd}{Local response trait diversity level}

\item{sp.subset}{sp.subset}

\item{patch.sp.num}{Number of species in a patch}

\item{warn.lst1}{First warning list}

\item{warn.lst2}{Second warning list}
}
\description{
Function to do local assignment
(function so that it can be invoked at multiple points in the code)
}

