% Generated by roxygen2 (4.0.1): do not edit by hand
\name{quant.reg}
\alias{quant.reg}
\title{Implement quantile regression}
\usage{
quant.reg(qr.out, RunID, d2, qr.vars, quantiles)
}
\arguments{
\item{qr.out}{File to which the results should be written. Needs to already exist with header information (created earlier in the code)}

\item{RunID}{ID to join to global scenario conditions}

\item{d2}{Input dataframe containing values}

\item{qr.vars}{A list of variables in the dataframe to use in quantile regression}

\item{quantiles}{The quantiles at which to do the evaluations}
}
\description{
Implement quantile regression
%% Re-enable this line when code is functional and properly integrated:
}

