% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{read.landscape.lyr}
\alias{read.landscape.lyr}
\title{Read landscape layer}
\usage{
read.landscape.lyr(landscape.dir, lbl, lnd.lbl, change.count, p)
}
\arguments{
\item{landscape.dir}{The path for this model run}

\item{lbl}{The name of the landscape layer being read in}

\item{lnd.lbl}{A scenario specific label identifying the landscape layers to be used.}

\item{change.count}{An indicator for the current environmental change step.}

\item{p}{Number of cells in the landscape}
}
\description{
Read in a landscape layer from a .csv.  Used when initially reading in a layer
and if c.type == "from.file".
}

