% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdemography.r
\name{Results.Outputs}
\alias{Results.Outputs}
\title{Results Outputs}
\description{
File outputs generated by the model (including some optional outputs
by requesting visual debugger data or matrix diagnostics.
}
\details{
Table summarizes how many files are produced
per model run (MR), scenario (SC), timestep (TS), cell (CL), species (SP),
environmental layer (EL), environmental conditions (EC). \cr
The file  production is multiplicative
(e.g., 100 timesteps x 2 species = 200 files) 
\tabular{lll}{
File type \tab Number of Files created \tab Link to File description \cr
Seeds in landscape \tab MR, SC, TS, SP \tab \link[spatialdemography]{stage.abundances} \cr
Juveniles in landscape \tab  MR, SC, TS, SP \tab \link[spatialdemography]{stage.abundances} \cr
Adults in landscape \tab  MR, SC, TS, SP \tab \link[spatialdemography]{stage.abundances} \cr
Local transition matrices \tab MR, SC, CL, SP, EC \tab \link[spatialdemography]{local.transition.matrices} \cr
Landscape-wide matrices \tab  MR, SC, SP, EC \tab \link[spatialdemography]{landscape.wide.matrix} \cr 
Local matrix summary info \tab MR, SC, SP, EC \tab \link[spatialdemography]{local.matrix.summary.info} \cr
Overall matrix diagnostics \tab MR, SC, EC \tab \link[spatialdemography]{overall.matrix.diagnostics} \cr
Landscape files \tab MR*, SC*, EL, EC; *can be reused \tab \link[spatialdemography]{landscape.files} \cr
Change lookup \tab MR, SC \tab \link[spatialdemography]{change.lookup} \cr
Cell occupancy \tab MR, SC \tab \link[spatialdemography]{cell.occupancy} \cr
Species statistics seeds \tab MR, SC \tab \link[spatialdemography]{species.statistics} \cr
Species statistics juveniles \tab MR, SC \tab \link[spatialdemography]{species.statistics} \cr
Species statistics adults \tab MR, SC \tab \link[spatialdemography]{species.statistics} \cr
Scenario Timing \tab MR, SC \tab \link[spatialdemography]{scenario.timing} \cr
Model Results \tab MR \tab \link[spatialdemography]{results.file} \cr
}
}

