% Generated by roxygen2 (4.0.1): do not edit by hand
\name{s.a.check}
\alias{s.a.check}
\title{Check species availability}
\usage{
s.a.check(my.set, my.subset, target.sp.num, target.rtd, draw.pool, my.draw,
  chk.type, warn.lst1, warn.lst2)
}
\arguments{
\item{my.set}{my.set}

\item{my.subset}{my.subset}

\item{target.sp.num}{target.sp.num}

\item{target.rtd}{target.rtd}

\item{draw.pool}{draw.pool}

\item{my.draw}{my.draw}

\item{chk.type}{chk.type}

\item{warn.lst1}{A list of warnings}

\item{warn.lst2}{A second list of warnings}
}
\description{
Check that the number of species available is greater than the number of
species needed based on response trait diversity
If not, re-draw until the criteria is reached or until the code gives up.
}

