% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save.seed}
\alias{save.seed}
\title{Save Seed}
\usage{
save.seed(run.id, seed.id)
}
\arguments{
\item{run.id}{Identifies which test run is being run}

\item{seed.id}{Identifies the location where the seed needs saving}
}
\description{
Simple function to create a seed token, that can later be assigned to .Random.seed
this will allow reproducibility of results, even when pieces of the code are moved around
The downside is that the code will slowly get cluttered with seed tokens, which at some point
will need to be replaced & the test code updated.
But this will forstall that day, and at least ensure that code that passed the test
is being used to replace the test code. And that reassures me greatly.
Thanks to Ben Bolker's answer on Stack Overflow:
 http://stackoverflow.com/questions/13997444/print-the-current-random-seed-so-that-i-can-enter-it-with-set-seed-later
The seed can be restored by doing load("SeedToken%s.RData"), where %s is the seed.id you want to access.
}

