% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scale.vec}
\alias{scale.cells.lst}
\alias{scale.vec}
\title{Setting the Scale of output results}
\description{
SpatialDemography has the ability to examine model results at multiple spatial scales.
The default settings scale.vec = c("landscape") and scale.cells.lst = "default"
examine model results at the scale of the entire landscape. The entries in
scale.vec are just labels, they do not actually serve any functional purpose.
scale.cells.lst actually determines which cells are evaluated, in the default
condition, a vector of all cells is included. This vector can just select
specific cells (i.e. a value of 1 for cells to be evaluated and 0 for cells
not to be evaluated. In addition, a new option has been added where only
cells of a particular environmental layer value can be evaluated. In this
case, scale.cells.lst should be set to "subset;X;Y" where X stands for a landscape
identifier, and Y stands for the value of that environmental layer to be considered.
For example, "subset;L;1" would look for the environmental layer with landscape.identifier of L
(typically landcover or land use for my simulations), and would only calculate
output results for a value of 1 (typically grassland cells in my simulations).
Unfortunately this cannot be used with evaluations of multiple scales.
But that limitation may change in the future.
}

