% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scenario.timing}
\alias{scenario.timing}
\title{Scenario Timing}
\description{
The Scenario timing file gives the time and overall elapsed time for
different model processes, as well as an overall model time.  The purpose of
this file is to know how long a model run took, and to identify rate-limiting
processes that can either be reimplemented to be faster in R, or implemented
in C or C++
}

