% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{scl.subset}
\alias{scl.subset}
\title{Create a subset for scale cells}
\usage{
scl.subset(scale.cells.lst, landscape.identifiers, landscape)
}
\arguments{
\item{scale.cells.lst}{In this case, a text field with three parts. The first
part is the word subset, the second part is the landscape.identifier to select
the landscape layer, and the third part is the target value from that landscape layer}

\item{landscape.identifiers}{A vector of landscape.identifiers}

\item{landscape}{The landscape consisting of environmental layers}
}
\description{
Select just a particular value of an environmental layer for evaluating
biodiversity metrics
}

