% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{sdr.lookup}
\alias{sdr.lookup}
\title{Look up Seed Dispersal Ratio}
\usage{
sdr.lookup(value, vec1, vec2)
}
\arguments{
\item{value}{a value}

\item{vec1}{a vector}

\item{vec2}{a vector}
}
\description{
Finds the seed dispersal ratio
}
\note{
This is the proportion of seeds dispersing to other cells
and is not actually a ratio (appears to be misnamed)
}

