% Generated by roxygen2 (4.0.1): do not edit by hand
\name{settings.file}
\alias{settings.file}
\title{Settings File}
\description{
This file contains settings for one or more model scenairios.  Each row
in this file corresponds to a scenario
}
\details{
\tabular{ll}{
Input                \tab Description \cr
Scenario             \tab A label for each scenario.  Needs to correspond &
                          be in the same order as the Scenario input in the
                          initial conditions file. \cr
GenerateVisualDebuggerData \tab An indicator variable.
                          When 0, additional diagnostics are not produced.
                          When 1, additional diagnostics are produced.\cr
GenerateNewSpecies   \tab When 1, new species are generated based on the
                          Species Instructions File
                          \link[spatialdemography]{sp.instr.file} and the
                          Species Response Instructions File
                          \link[spatialdemography]{sp.resp.instr.file}.
                          Otherwise, a pointer should be given to indicate
                          which existing Species File
                          (\link[spatialdemography]{sp.file}) to use
                          (the pointer is a short piece of code in the
                          Species file name that follows species_file_ and
                          precedes any ending (e.g., date).\cr
GenerateNewLandscape \tab When 1, new landscape files are generated based on
                          the Environmental Layers File
                          (\link[spatialdemography]{env.file}).\cr
ignore.rtd           \tab (optional) An indicator variable for whether
                          response trait diversity should be taken into
                          account when assigning species.
                          If 1, specified response trait diversity levels
                          will be met (or an error will be given if they
                          cannot be met),
                          if 0, the model will proceed without respect to
                          those inputs. Not applicable when species locations
                          are read in from file. \cr
initial.n            \tab (optional, if column is absent, defaults to a value
                          of 1.) An indicator for how initial species numbers
                          should be assigned.  If absent or 1, 10 times as
                          many juveniles and 100 times as many seeds will be
                          produced (relative to adult numbers). If 2, a
                          starting biomass is specified in the
                          \link[spatialdemography]{initial.conditions.file}
                          and this biomass is divided evenly among all three
                          life stages.  Each life stage starts with a number
                          of individuals equal to 1/3 starting biomass
                          divided by the biomass of an individual
                          (floor division). \cr
loc.extraction       \tab (optional) Indicates how extraction from file
                          should proceed. Extraction types are: 1. Read in
                          all life stages from file. 2. Read in locations
                          (presence) from file, but set up abundances based
                          on settings in initial.n. 3. Read in adult
                          locations from file, and set up juveniles and seeds
                          based on settings in initial.n. \cr
timestep.extraction  \tab The time step to be extracted from file \cr
include.copula       \tab (optional) An indicator for whether generation via
                          a copula is desired.  Number should indicate the
                          number of different copulas desired. \cr
N.sim.dispersal.table \tab The number of random draws to be used for
                          obtaining simu-lated dispersal probabilities.\cr
RunSimulation        \tab When 1, this indicates that the simulation portion
                          of the model should be run, otherwise the
                          simulation is not con-ducted.\cr
RunMatrixDiagnostics \tab When 1, this indicates that additional matrix
                          diagnostics should be computed and written to file,
                          otherwise these are not computed.\cr
Write.Timefile       \tab When 0 or absent, no timing is done.
                          When 1, the total scenario time is written to file.
                          When 2, total scenario time and run time details
                          are output to file.\cr
}
}

