% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setup.abund}
\alias{setup.abund}
\title{Set up species abundances in the landscape}
\usage{
setup.abund(n0.lst, spe, species.locs, S, initial.n, ic, SpTraits)
}
\arguments{
\item{n0.lst}{A list of abundances for each stage for each species}

\item{spe}{The total number of species in the species pool}

\item{species.locs}{The locations of each species in the landscape}

\item{S}{The number of life stages in the model}

\item{initial.n}{A setting for initial number of species desired}

\item{ic}{The initial conditions read in from the Initial Conditions File}

\item{SpTraits}{The species traits, read in from the species file}
}
\description{
Set up species abundances in the landscape
}

