% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setup.copula}
\alias{setup.copula}
\title{Function to set up copula}
\usage{
setup.copula(my.env, MaxTime, p, include.copula, run.path, env.lbl, s.lbl)
}
\arguments{
\item{my.env}{Information about the environmental layers}

\item{MaxTime}{The number of timesteps the model will be run for}

\item{p}{The number of cells in a square landscape}

\item{include.copula}{The number of copulas to create}

\item{run.path}{The path for the analysis}

\item{env.lbl}{A label}

\item{s.lbl}{A label}
}
\description{
This function sets up a copula for the model. A copula is a set of variables
generated with a certain level of correlation.
}

