% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setup.env.change.mag}
\alias{setup.env.change.mag}
\title{Setup env.change.mag}
\usage{
setup.env.change.mag(env.change.freq, env.change.type, in.vals = NA,
  in.transitions = NA, in.delta.transitions = NA, trans.type = NA)
}
\arguments{
\item{env.change.freq}{The frequency of environmental change. Mainly included to ensure that it is not 0}

\item{env.change.type}{The type of environmental change. For more info, see \link[spatialdemography]{env.change.type}}

\item{in.vals}{Input values excluding transition rates or changes in transition rates (changes to mean or sd SHOULD be included when appropriate for the change type)}

\item{in.transitions}{Only applies to Markov approaches. This should be a matrix of transitions, starting with the lowest cell value to each other cell value (1st row), then an additional row for each cell value.}

\item{in.delta.transitions}{Only applies to changing Markov approaches. This should indicate the change in transition probability.}

\item{trans.type}{Acceptable values: "ind", "cum". Only applies to Markov approaches. Indicates whether transition probabilities are given for each transition ("ind"), or cumulatively to include prior transitions ("cum").}
}
\description{
This is a helper function to facilitate setting up the env.change.mag field in
the Enviornmental_layers file. This may not be relevant for many environmental change types.
%% Add example for Markov version?
}

