% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setup.env.info}
\alias{setup.env.info}
\title{Set up env.info variable}
\usage{
setup.env.info(env.type, my.env, index)
}
\arguments{
\item{env.type}{An indicator for what type of environmental values are present (e.g., cover, normal random)}

\item{my.env}{Information about the environmental layers}

\item{index}{An index}
}
\description{
Function to set up env.info and reduce code redundancy and increase
streamlining
}

