% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{setup.inputs}
\alias{setup.inputs}
\title{Set up species and environmental inputs for stochastic plants simulation}
\usage{
setup.inputs(in.info, in.lhs, in.lbl = "Species")
}
\arguments{
\item{in.info}{A file containing data to be sampled according to the hypercube design}

\item{in.lhs}{A latin hypercube sampling matrix}

\item{in.lbl}{A label for the dataframe. Can be "Species" or "Model.Run"}
}
\value{
returns the created species file read into R.
}
\description{
For stochastic plants simulation.  Creates according to Gerstenlauer et al. (in prep).
}

