% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{setup.locations}
\alias{setup.locations}
\title{Set up species locations in the landscape}
\usage{
setup.locations(n0.lst, spe, S, p, ic, settings, locations.file, SpTraits,
  rtd.c, run.path)
}
\arguments{
\item{n0.lst}{An initial abundance vector containing abundances of every species at every stage for every cell}

\item{spe}{The total number of species in the landscape}

\item{S}{The number of stages in the matrix model}

\item{p}{The number of cells in a square landscape}

\item{ic}{The initial conditions read in from the initial conditions file}

\item{settings}{The initial settings, read in from the initial settings file}

\item{locations.file}{A file giving the species locations. If no such file exists, this parameter should be set to "none".}

\item{SpTraits}{The Species traits, read in from the species file}

\item{rtd.c}{An indicator identifiying response traits. Currently broken}

\item{run.path}{The path for the analysis.}
}
\description{
Function to implement one of several options for species assignment
}

