% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{setup.results}
\alias{setup.results}
\title{Create results file}
\usage{
setup.results(ResultsFile, out.metrics, scale.vec, do.cwm = 0)
}
\arguments{
\item{ResultsFile}{The file that results will be written to.}

\item{out.metrics}{The fields to be calculated}

\item{scale.vec}{An indicator vector}

\item{do.cwm}{An indicator for whether community weighted means should be calculated. (Needs testing, may be broken)}
}
\description{
Set up the output results file and results dataframe
}

