% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setup.results.v2}
\alias{setup.results.v2}
\title{Create results file in a new format}
\usage{
setup.results.v2(ResultsFile, out.metrics, nrow.results)
}
\arguments{
\item{ResultsFile}{The file that results will be written to.}

\item{out.metrics}{The fields to be calculated}

\item{nrow.results}{Number of rows in the final results file}
}
\description{
Set up the output results file and results dataframe, with a row for each content type, and a row for initial, final and change
}

