% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{setup.rtd}
\alias{setup.rtd}
\title{Estimate Response Trait Diversity}
\usage{
setup.rtd(SpTraits, rtd.c = 0)
}
\arguments{
\item{SpTraits}{A dataframe with species information}

\item{rtd.c}{Currently non-functional input}
}
\value{
The SpTraits dataframe with a unique code assigned to each response
type
}
\description{
This function takes a dataframe with species information, and calculates
the response trait diversity present (giving each species a response trait
diversity number that corresponds to a unique combination.  Some formatting
caveats apply. Dispersal traits need to be in columns 11 - 13, and any
remaining response traits in columns 14+
}

