% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Simulation}
\alias{Simulation}
\title{Function to run the simulation portion of SpatialDemography}
\usage{
Simulation(Model.Name, ResultsFile, vdb.data, timefile, write.timefile,
  run.times, run.lbl, start.time, run.path, DispPath, outpath.base, num.sim, S,
  extent, p, landscape, landscape.identifiers, distances, settings, ic, rtd.c,
  locations.file, landscape.dir, SpTraits, tot.sp.num, my.env, env.c.freq,
  is.change, change.count, env.lbl, s.lbl, lnd.lbl, competitiontype, microsites,
  Results, out.metrics, scale.vec, timepoint.vec, scale.cells.lst, resolution,
  log.trans, MaxTime, invasion, num.invaders, cells.to.invade, repro.proportion,
  K_g, multi.species.K, edge.type, do.simulation, do.diagnostics, testing)
}
\arguments{
\item{Model.Name}{The name of the model run}

\item{ResultsFile}{The file that results will be written to.}

\item{vdb.data}{An indicator for whether or not to create visual debugger data}

\item{timefile}{Timing results are written to this file.}

\item{write.timefile}{An indicator for whether the time file should be output.}

\item{run.times}{A list of intermediate times during the model run}

\item{run.lbl}{A list of labels for each of the intermediate times}

\item{start.time}{The starting time for the model}

\item{run.path}{The path for the model run}

\item{DispPath}{The path for the dispersal tables}

\item{outpath.base}{A base path for outputs}

\item{num.sim}{The number of simulations to run}

\item{S}{The number of life stages in the model (MUST BE 4)}

\item{extent}{The length of one side of the square landscape}

\item{p}{The number of cells in the landscape (extent ^ 2)}

\item{landscape}{A list object containing the environmental layer values for each cell in the landscape}

\item{landscape.identifiers}{A list of single letter identifiers for each environmental layer}

\item{distances}{The distances between every pair of cells}

\item{settings}{A dataframe containing the information from the settings file}

\item{ic}{A dataframe containing the information from the initial conditions file}

\item{rtd.c}{An indicator for how to handle response trait diversity. Currently non-functional, use the hard-wired default}

\item{locations.file}{A file indicating species locations. May not be applicable in many circumstances.}

\item{landscape.dir}{A directory pointing to where the landscape files should be written.}

\item{SpTraits}{Species trait data read in from the species file}

\item{tot.sp.num}{The number of species in the species pool (=spe)}

\item{my.env}{Information about the environmental layers}

\item{env.c.freq}{The frequency of environmental change}

\item{is.change}{An indicator for whether or not change will occur in the model (for computational efficiency)}

\item{change.count}{An indicator for the current environmental change step.}

\item{env.lbl}{A vector of labels containing the landscape layer names}

\item{s.lbl}{Another label}

\item{lnd.lbl}{Another label}

\item{competitiontype}{An indicator for what type of competition should be implemented}

\item{microsites}{The number of microsites in the model}

\item{Results}{A dataframe to hold the results from the model}

\item{out.metrics}{A list of metrics to be calculated for the model}

\item{scale.vec}{A list of spatial scales at which to evaluate the model}

\item{timepoint.vec}{A list of timepoints at which to calculate the metrics given in out.metrics}

\item{scale.cells.lst}{A list of cells corresponding to each spatial scale}

\item{resolution}{The resolution to use for rounding (for calculating UTC values)}

\item{log.trans}{Whether or not to log-transform the data. Currently not recommended, as this functionality is untested in multirich package}

\item{MaxTime}{The maximum number of timesteps to run the model for.}

\item{invasion}{How frequently invasion occurs}

\item{num.invaders}{The number of species to have invade}

\item{cells.to.invade}{The number of cells each species will invade}

\item{repro.proportion}{The proportion of the species reproductive potential with which to invade}

\item{K_g}{The carrying capacity in grams}

\item{multi.species.K}{An indicator for whether or not to include a multispecies carrying capacity}

\item{edge.type}{The edge type. Currently only ABSORBING and TORUS are supported}

\item{do.simulation}{An indicator for whether or not to run the simulation}

\item{do.diagnostics}{An indicator for whether or not to compute diagnostics}

\item{testing}{An indicator for whether or not a testing run is being conducted}
}
\description{
This function does the model setup, matrix diagonstics, and simulation portion of spatialdemography.
%DD% ADD INPUT DESCRIPTIONS %%s.lbl is only really required for reading in changed landscape values from file.
%% disp.pdf,eigen.pdf,rtd.c
}
\author{
Alexander "Sasha" Keyel & Jakob Gerstenlauer
}

