% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sp.instr.file}
\alias{sp.instr.file}
\title{Species Instructions File}
\description{
Optional input file giving instructions for generation of species base vital
rates. Each vital rate should be a row entry. For each vital rate, three
columns need to be filled in: Function, parameter1, and parameter2 see
\link[spatialdemography]{distribution.functions} for more information. Either
this file and the \link[spatialdemography]{sp.resp.instr.file}
OR the \link[spatialdemography]{sp.file} are required.
}
\details{
\tabular{ll}{
Input            \tab Description \cr
p01              \tab Transition rate from stage 0 to stage 1.\cr
p02              \tab Transition rate from stage 0 to stage 2.\cr
p11              \tab Survival rate of stage 1 (proportion of stage 1
                      individuals that remains in stage 1).\cr
p12              \tab Transition rate from stage 1 to stage 2.\cr
p22              \tab Survival rate of stage 2 (proportion of stage 2
                      individuals that remains in stage 2).\cr
p23              \tab Transition rate from stage 2 to stage 3.\cr
p30              \tab Number of stage 0 individuals produced by stage 3.\cr
p32              \tab Number of stage 2 individuals produced by stage 3.\cr
p33              \tab Survival rate of stage 3 (proportion of stage 3
                      individuals that remain in stage 3).\cr
biomass.adult    \tab The species' adult biomass (g).\cr
biomass.juv      \tab The species juvenile biomass (g), only needed if
                      assigning initial numbers by biomass. \cr
biomass.seed     \tab The species' seed mass (g), only needed if assigning
                      initial numbers by biomass. \cr
dispersalfunction \tab A number indicating which dispersal function to use.
                      Log-normal is indi-cated by 1.\cr
disppar1         \tab The first input parameter for the dispersal function.
                      For log-normal, this is the mean dispersal distance,
                      in cells (not log transformed).\cr
disppar2         \tab The second input parameter for the dispersal function.
                      For log-normal, this is the dispersal standard
                      deviation in cells (log-transformed).\cr
}
}

