% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sp.param.file}
\alias{sp.param.file}
\title{Species Parameters File}
\description{
The file contains three headings: Trait, Min, and Max.  Trait lists traits,
while min and max give minimum and maximum values to sample from.  When a
fixed value is desired, min and max should be equal.
See details for description of traits that should be included.
}
\details{
\itemize{
\item seed.disp.ratio The proportion of seeds dispersing out of the cell.
      %% This parameter is misnamed, should be proportion dispersing.
\item adult.longevity Adult survival %% This parameter is also misnamed.
\item g Proportion of reproduction allocated to clonal reproduction
\item w Probability of a seed germinating
\item biomass.seed Biomass of a seed
\item biomass.adult Biomass of an adult
\item biomass.clone Biomass of a clone
}
}

