% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{sp.vals}
\alias{sp.vals}
\title{Species Values Input}
\description{
Input species file for calculate traits %DD% Add documentation
}
\details{
\itemize{
\item biomass.seed Biomass of the species' seed (= seed mass).  Used in thinning & reproductive trade-off\cr
\item biomass.adult Biomass of the adult.  Used in thinning law \cr
\item  biomass.clone Biomass of clones produced.  Used in thinning law & reproductive trade-off \cr
\item  g The proportion of reproductive effort allocated to producing clones \cr
\item  seed.disp.ratio The proportion of seeds dispersing to other cells in the landscape \cr
\item  w The proportion of seeds germinating at each time step \cr
\item  seed.mortality The proportion of seeds that die \cr
\item  adult.longevity The proportion of adults that survive each time step \cr
\item  Ss Stochastic seedling survival #DD# needs more documentation \cr
\item  k2 A parameter for the Weibull function that determines reproductive allocation \cr
\item  fertility A multiplier that determines how many seeds/clones are produced for a given reproductive allocation \cr
\item  sp.out.file The path and file where the new species file will be written (or of an existing file to be appended to) \cr
\item  do.append Whether to create a new file (F) or to append to an existing file (T) \cr
\item  n.seeds.o An option to override seed production (independent of the trade-off).  Both n.seeds.o and n.clones.o must be non-null to do the override. \cr
\item  n.clones.o An ooption to override clone production.  Both this and n.seeds.o must be non-null to override the trade-offs. \cr
\item  d12.o An option to override the thinning portion of the seed -> juvenile transition rate.  All d**.o must be non-null for the override to take place. \cr
\item  d23.o An option to override the thinning portion of the juvenile -> adult transition rate.  All d**.o must be non-null for the override to take place. \cr
\item  dc2.o An option to override the thinning portion of the clone -> juvenile transition rate.  All d**.o must be non-null for the override to take place. \cr
}
}

