% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SpatialDemography}
\alias{SpatialDemography}
\title{SpatialDemography = Main function for running the spatially explicit,
matrix-based, metacommunity model}
\usage{
SpatialDemography(scn, s.lbl, file.ending, DispPath, run.path, opath,
  ResultsFile, initial.conditions.file, settings.file, env.file, spfile,
  sp.instr.file = "none", sp.resp.instr.file = "none",
  landscape.dir = "default", locations.file = "none",
  scale.vec = c("landscape"), scale.cells.lst = "default",
  out.metrics = c("Sp.Rich", "Beta.Div", "Biomass", "FTD.UTC", "RTD.UTC"),
  testing = F)
}
\arguments{
\item{scn}{Indicator variable to denote scenario and to extract appropriate
row from initial.conditions.file and settings.file.}

\item{s.lbl}{A label used in the construction of the Species file and new
landscape files.}

\item{file.ending}{A text string that matches the end of the species file
(either to be gener-ated or that already exists).}

\item{DispPath}{The path of the dispersal tables (if existing) or the path
where dispersal tables should be created.}

\item{run.path}{The main path for the specific overall model run.}

\item{opath}{The path for model outputs.}

\item{ResultsFile}{A file to be created containing the model results.}

\item{initial.conditions.file}{A file to specify the initial model
conditions. See \link[spatialdemography]{initial.conditions.file} for details.}

\item{settings.file}{A file to specify overall model settings. See
\link[spatialdemography]{settings.file} for details.}

\item{env.file}{A file to describe environmental layers and how they change.
See \link[spatialdemography]{env.file} for details.}

\item{spfile}{A file specifying species traits.
See \link[spatialdemography]{sp.file} for details.}

\item{sp.instr.file}{A file to describe how generation of species' base
vital rates is to occur (Optional - not needed if using pre-existing species
file). See \link[spatialdemography]{sp.instr.file} for details.}

\item{sp.resp.instr.file}{A file to describe how generation of species'
response traits is to occur (Optional - not needed if using pre-existing
species file). See \link[spatialdemography]{sp.resp.instr.file} for details.}

\item{landscape.dir}{(not used if generating new landscapes) The directory
where the landscape files may be found or a list object containing the
landscape information. See \link[spatialdemography]{landscape.files} for
more information on the files, and see
\link[spatialdemography]{landscape.object} for information on using a
landscape generated in R.}

\item{locations.file}{(should be 'none' if spatialdemography is assigning
species to the landscape) A file containing initial locations of species.
See (\link[spatialdemography]{locations.file}) for more information
about the species locations file itself.}

\item{scale.vec}{A list of spatial scales at which to evaluate the model.
See \link[spatialdemography]{scale.vec} for more details.}

\item{scale.cells.lst}{A list of cells from the landscape included in each of
the above spatial scales.
See \link[spatialdemography]{scale.vec} for more details.}

\item{out.metrics}{Metrics to calculate for the model run.
See \link[spatialdemography]{out.metrics} for details.}

\item{testing}{An indicator for whether test files are being run. Defaults to F.}
}
\value{
No direct results are returned from the function.  However, the
function writes several files.
See \link[spatialdemography]{Results.Outputs} for a summary.
}
\description{
This function implements a spatially explicit matrix model based on input
options. Inputs are pointers to files, which contain additional inputs used
by the function to run the model.
}
\details{
Details for each of the inputs in the input files: Please see Overview,
Design concepts, Details (ODD, Grimm et al. 2006, 2010) description for
SpatialDemography %% Figure out how to link to this here
}
\author{
A.C. Keyel
}
\references{
Keyel, A. C., J. L. K. Gerstenlauer, and K. Wiegand. (in prep)
spatialdemography: an R package for implementing spatially explicit, matrix-based,
metacommunity models.
}

