% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{spatialdemography_package}
\alias{spatialdemography_package}
\alias{spatialdemography_package-package}
\title{MetaComeSE (package)}
\description{
Package to run a spatially explicit, matrix-based metacommunity model
}
\details{
\tabular{ll}{ %% Note tabular{ll} is tabular{lowercase(LL)} not tabular{11}
Package: \tab spatialdemography\cr
Type: \tab Package\cr
Version: \tab 0.10.3\cr
Date: \tab 2015-04-02\cr
License: \tab GPL-2 (or later)\cr
}
}
\author{
Alexander "Sasha" Keyel & Jakob L.K. Gerstenlauer\cr
Maintainer: Sasha Keyel <skeyel@gmail.com>
}
\references{
Keyel, A. C., J. L. K. Gerstenlauer, and K. Wiegand. (in prep)
spatialdemography: an R package for implementing spatially explicit, matrix-based,
metacommunity models.
}
\seealso{
\link[spatialdemography]{SpatialDemography}
}
\keyword{package}

