% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdemography.r
\name{species.statistics}
\alias{species.statistics}
\title{Species Statistics}
\description{
Summary statistics by time step for each species in the landscape.
Time steps correspond to rows, and species correspond to columns (this is
somewhat non-intuitive and confusing in the file)
}

