% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{spp.subset}
\alias{spp.subset}
\title{Species subset}
\usage{
spp.subset(my.subset, target.rtd, target.sp.num, warn.lst1, warn.lst2)
}
\arguments{
\item{my.subset}{A subset}

\item{target.rtd}{a response trait diversity target}

\item{target.sp.num}{The target species number}

\item{warn.lst1}{A list of warnings}

\item{warn.lst2}{A second list of warnings}
}
\description{
Function to reduce species number from a regional species pool to the target
species number.
}

