% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdemography.r
\name{stage.abundances}
\alias{adult.abundances}
\alias{juvenile.abundances}
\alias{seed.abundances}
\alias{stage.abundances}
\title{Stage Abundances}
\description{
These files are in a folder labeled *_matrices, where * corresponds to the 
lifestage of interest.  This gives the abundance of the lifestage of interest
in every cell of the landscape.  There is one file per species per timestep,
for each scenario. \cr
File names are Stage_species_timestep.csv
}

