% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{stoch.plant.setup}
\alias{stoch.plant.setup}
\title{Stochastic Plants Setup}
\usage{
stoch.plant.setup(SampleSize, ModelRuns, sp.param.file, env.param.file,
  do.seed.clone.override = 0, do.thinning.override = 0, NumSVars = 7,
  NumEVars = 7)
}
\arguments{
\item{SampleSize}{Number of species to generate}

\item{ModelRuns}{Number of scenarios to generate}

\item{sp.param.file}{A file containing species parameters.  See \link[spatialdemography]{sp.param.file} for details.}

\item{env.param.file}{A file containing environmental parameters.  See \link[spatialdemography]{env.param.file} for details.}

\item{do.seed.clone.override}{0 indicates that seed and clone number should be treated as dependent traits
1 allows these traits to be specified as independent traits.}

\item{do.thinning.override}{0 indicates that some demographi transitions should be dependent traits
based on the thinning law.  1 indicates that these parameters may be independently speciefied.}

\item{NumSVars}{Number of species variables in sp.param.file input.  Default is 7.}

\item{NumEVars}{Number of environmental variables in env.param.file.  Default is 7.}
}
\description{
Set up latin hypercube sampling for plant traits used in stochastic plants simulation 
%% #DD# add documentation for env.param.file and sp.param.file
}

