% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test.setup}
\alias{test.setup}
\title{Test Setup}
\usage{
test.setup(base.dir, run.name)
}
\arguments{
\item{base.dir}{a base directory containing the relevant files}

\item{run.name}{the run name and output directory}
}
\value{
Returns a list of the five files that are inputs to SpatialDemography
}
\description{
This function helps set up the file structure for the model accuracy/implementation test
}

