% Generated by roxygen2 (4.0.1): do not edit by hand
\name{torus.edge}
\alias{torus.edge}
\title{Torus Edge}
\usage{
torus.edge(i, p, BaseProbabilities)
}
\arguments{
\item{i}{The location of the specific cell}

\item{p}{The number of cells in a square landscape}

\item{BaseProbabilities}{Base dispersal probabilities}
}
\description{
Implement dispersal probabilities,
where individuals wrap around to the other edge of the world
}

