% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validity.check}
\alias{validity.check}
\title{Validity Check}
\usage{
validity.check(x)
}
\arguments{
\item{x}{An input probability to be checked}
}
\description{
Check that probabilities of environmental change are actually probabilities
(and if not, adjust them)
}

