% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{write.env.lyr}
\alias{write.env.lyr}
\title{Write environmental layer to file}
\usage{
write.env.lyr(in.lyr, change.count, extent, outfile, create.new = 0)
}
\arguments{
\item{in.lyr}{The input layer values}

\item{change.count}{An indicator for the current environmental change step.}

\item{extent}{One sided length of the landscape}

\item{outfile}{File to be created}

\item{create.new}{(default to 0) An indicator for whether to write header info & start a new file}
}
\description{
Write an environmental layer to file (for visual debugging purposes)
Input is a landscape as a vector.  The landscape can be visualized by putting it into a matrix format (nrow = extent) and using the byrows = T option
}

