% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.sp.data}
\alias{write.sp.data}
\title{Write species locations to file}
\usage{
write.sp.data(LifeStage, sp, RunTime, in.sp.dat, extent, outfile)
}
\arguments{
\item{LifeStage}{The life stage of interest}

\item{sp}{the current species}

\item{RunTime}{The current timestep}

\item{in.sp.dat}{in.sp.dat}

\item{extent}{The one-sided length of the landscape}

\item{outfile}{the file to be written to}
}
\description{
Write the locations of a species to a file
Input is species abundances as a vector.  The species abundances on the landscape can be visualized by putting it into a matrix format (nrow = extent) and using the byrows = T option
}

