% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.sp.hdr.data}
\alias{write.sp.hdr.data}
\title{Write header for species data}
\usage{
write.sp.hdr.data(outfile, p)
}
\arguments{
\item{outfile}{The file to be written to}

\item{p}{The number of cells in a square landscape}
}
\description{
Like the title says.
}

