% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{wrt}
\alias{wrt}
\title{Write table for timing function}
\usage{
wrt(in.vec, timefile, do.append = TRUE)
}
\arguments{
\item{in.vec}{the vector to be written to table}

\item{timefile}{the file to write to.}

\item{do.append}{Whether or not to append to an existing file}
}
\description{
Simple version to get write.table to behave properly
}

