% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{Ztrans}
\alias{Ztrans}
\title{Z transform}
\usage{
Ztrans(x)
}
\arguments{
\item{x}{an input variable}
}
\description{
Function for z transforming input variables. Used by qr.bqr
}
\author{
Jakob L.K. Gerstenlauer
}

