% Generated by roxygen2 (4.0.1): do not edit by hand
\name{absorbing.edge}
\alias{absorbing.edge}
\title{Absorbing edge}
\usage{
absorbing.edge(i, p, BaseProbabilities)
}
\arguments{
\item{i}{The location of the specific cell}

\item{p}{The number of cells in a square landscape}

\item{BaseProbabilities}{Base dispersal probabilities}
}
\description{
Implement the dispersal probabilities, with species able to fall off the edge
of the world
}

