% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{change.check}
\alias{change.check}
\title{Check for environmental change}
\usage{
change.check(env.c.freq)
}
\arguments{
\item{env.c.freq}{Frequency of environmental change}
}
\description{
Test if there will ever be environmental changes in a model run
}

