% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{CreatePermutationMatrix}
\alias{CreatePermutationMatrix}
\title{Create Permutation Matrix}
\usage{
CreatePermutationMatrix(p, S)
}
\arguments{
\item{p}{The number of cells in a square landscape}

\item{S}{The number of stages in the matrix model}
}
\description{
Creates the vec-permutation matrix for use in running the matrix model
}
\author{
Jakob Gerstenlauer
}

