% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.cov}
\alias{get.cov}
\title{Calculate modified coefficient of varation}
\usage{
get.cov(d2)
}
\arguments{
\item{d2}{A data frame}
}
\description{
Calcualte modified coefficient of variation %% #DD# Add description for rationale here
}
\author{
Jakob Gerstenlauer, converted to function by Sasha Keyel
}

