% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{get.vrs}
\alias{get.vrs}
\title{Get vital rates}
\usage{
get.vrs(vrs.out, vrs, sp.base, tot.sp.num)
}
\arguments{
\item{vrs.out}{A vector to contain the final vital rates}

\item{vrs}{A vector of vital rates}

\item{sp.base}{Information about the vital rates and their distribution for species generation}

\item{tot.sp.num}{Total number of species (== spe)}
}
\description{
Function to loop through vital rates and get values
}

