% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{matrix.as.vector}
\alias{matrix.as.vector}
\title{matrix.as.vector}
\usage{
matrix.as.vector(in.mat)
}
\arguments{
\item{in.mat}{A matrix}
}
\description{
Converts a matrix to a vector, but does so by rows (I don't know of a pre-existing option for this)
}

