% Generated by roxygen2 (4.0.1): do not edit by hand
\name{old.style.locs}
\alias{old.style.locs}
\title{Original format for locations file}
\usage{
old.style.locs(settings, run.path)
}
\arguments{
\item{settings}{The settings from the read-in settings file}

\item{run.path}{The path for the analysis run}
}
\description{
Preserves backwards compatibility, will probably be deprecated at some point.
}

