% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{prep.output}
\alias{prep.output}
\title{Prep diagnostic output}
\usage{
prep.output(in.spar.mat, change.count, sp, cell = "")
}
\arguments{
\item{in.spar.mat}{in.spar.mat}

\item{change.count}{An indicator for the current environmental change step.}

\item{sp}{The species of interest (numeric)}

\item{cell}{A cell label}
}
\description{
Reformats the matrix into a vector, and adds desired prefix information
}

