% Generated by roxygen2 (4.0.1): do not edit by hand
\name{recode}
\alias{recode}
\title{Convert abundance to occupancy}
\usage{
recode(x)
}
\arguments{
\item{x}{Input a number >= 0.}
}
\value{
Outputs a recoded number: 0 (absent) or 1 (present).
}
\description{
Abundances > 0 are coded as present (1),
abundances == 0 are coded as absent (0).
}

