% Generated by roxygen2 (4.0.1): do not edit by hand
\name{target.recode}
\alias{target.recode}
\title{Target recode}
\usage{
target.recode(x, target)
}
\arguments{
\item{x}{An input number}

\item{target}{the target to be matched}
}
\description{
Recode such that all target values are 1 and all other values are 0
}

