% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{validity.check}
\alias{validity.check}
\title{Validity Check}
\usage{
validity.check(x)
}
\arguments{
\item{x}{An input probability to be checked}
}
\description{
Check that probabilities of environmental change are actually probabilities
(and if not, adjust them)
}

