% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdhelper.r
\name{write.sp.hdr.data}
\alias{write.sp.hdr.data}
\title{Write header for species data}
\usage{
write.sp.hdr.data(outfile, p)
}
\arguments{
\item{outfile}{The file to be written to}

\item{p}{The number of cells in a square landscape}
}
\description{
Like the title says.
}

